isolating seeds_helper.rb, using block form
