+++ test/api/external_resources_controller_test.rb
@@ -1,5 +1,5 @@
 require 'api_test_helper'
-require 'seeds_helper'
+require 'user_tokens_helper'
 
 class ExternalResourcesControllerTest < ActionDispatch::IntegrationTest
   include Devise::Test::IntegrationHelpers
